Add SignOn plus various fixes (#84)
* config-libraries.go: Add SignOn

* transformtypedefs.go: Add workaround for broken type

* main.go: Remove unused parameter

* config-allowlist.go: Update blocked types

* emitcabi.go: Update type mappings and remove unused parameter

* emitzig.go: Update type mappings, documentation comments, and special cases

* build.zig: Add SignOn and update cross-platform support

* README: Update installation instructions and usage of build option

* Update bindings and wrappers

+++ include/opengl/libqopenglfunctions_1_5.zig
@@ -639,22 +639,22 @@ pub const qopenglfunctions_1_5 = struct {
 
     /// [Qt documentation](https://doc.qt.io/qt-6/qopenglfunctions-1-5.html#glGetBufferSubData)
     ///
-    /// ``` self: QtC.QOpenGLFunctions_1_5, target: u32, offset: i64, size: i64, data: *void ```
-    pub fn GlGetBufferSubData(self: ?*anyopaque, target: u32, offset: i64, size: i64, data: *void) void {
+    /// ``` self: QtC.QOpenGLFunctions_1_5, target: u32, offset: isize, size: isize, data: *void ```
+    pub fn GlGetBufferSubData(self: ?*anyopaque, target: u32, offset: isize, size: isize, data: *void) void {
         qtc.QOpenGLFunctions_1_5_GlGetBufferSubData(@ptrCast(self), @intCast(target), @intCast(offset), @intCast(size), data);
     }
 
     /// [Qt documentation](https://doc.qt.io/qt-6/qopenglfunctions-1-5.html#glBufferSubData)
     ///
-    /// ``` self: QtC.QOpenGLFunctions_1_5, target: u32, offset: i64, size: i64, data: *void ```
-    pub fn GlBufferSubData(self: ?*anyopaque, target: u32, offset: i64, size: i64, data: *const void) void {
+    /// ``` self: QtC.QOpenGLFunctions_1_5, target: u32, offset: isize, size: isize, data: *void ```
+    pub fn GlBufferSubData(self: ?*anyopaque, target: u32, offset: isize, size: isize, data: *const void) void {
         qtc.QOpenGLFunctions_1_5_GlBufferSubData(@ptrCast(self), @intCast(target), @intCast(offset), @intCast(size), data);
     }
 
     /// [Qt documentation](https://doc.qt.io/qt-6/qopenglfunctions-1-5.html#glBufferData)
     ///
-    /// ``` self: QtC.QOpenGLFunctions_1_5, target: u32, size: i64, data: *void, usage: u32 ```
-    pub fn GlBufferData(self: ?*anyopaque, target: u32, size: i64, data: *const void, usage: u32) void {
+    /// ``` self: QtC.QOpenGLFunctions_1_5, target: u32, size: isize, data: *void, usage: u32 ```
+    pub fn GlBufferData(self: ?*anyopaque, target: u32, size: isize, data: *const void, usage: u32) void {
         qtc.QOpenGLFunctions_1_5_GlBufferData(@ptrCast(self), @intCast(target), @intCast(size), data, @intCast(usage));
     }
